+++ src/main/java/me/lokka30/levelledmobs/misc/FileLoader.java
@@ -16,9 +16,9 @@ import java.io.FileInputStream;
  */
 public final class FileLoader {
 
-    public static final int SETTINGS_FILE_VERSION = 28; // Last changed: b289
-    public static final int MESSAGES_FILE_VERSION = 2; // Last changed: v2.3.0
-    public static final int CUSTOMDROPS_FILE_VERSION = 7; // Last changed: 2.4.0 b328
+    public static final int SETTINGS_FILE_VERSION = 29; // Last changed: v2.6.0 b274
+    public static final int MESSAGES_FILE_VERSION = 3; // Last changed: v2.6.0 b274
+    public static final int CUSTOMDROPS_FILE_VERSION = 8; // Last changed: v2.6.0 b274
     public static final int CUSTOMCOMMANDS_FILE_VERSION = 1;
 
     private FileLoader() {